comments and fixes to homing missiles

+++ homing.h
@@ -4,19 +4,26 @@
 #include <QGraphicsPixmapItem>
 #include "player.h"
 
+
+//objects that move towards the player
 class Homing : public QGraphicsPixmapItem
 {
 public:
+    //constructor, creates the item outside of the screen and targets the player
     Homing(Player* player);
+    //moves the item
     void move(Player* player);
+    //returns x coordinate in int
     int getx();
+    //-||- for y
     int gety();
+
 
 private:
     float xlocation;
     float ylocation;
     float angle;
-    int accel;
+    float accel;
     float xspeed;
     float yspeed;
 };
